requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
diff-match-patch==20241021
Django==5.1.4
django-cors-headers==4.6.0
django-import-export==4.3.3
django-unfold==0.43.0
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
isort==5.13.2
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tablib==3.7.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
